+++ src/Inventory/Items/ElectronicsItem.java
@@ -3,7 +3,7 @@ package Inventory.Items;
 public class ElectronicsItem extends AbstractItem {
 
 
-    public ElectronicsItem(String name, String category, double price, double quantity) {
+    public ElectronicsItem(String name, String category, double price, int quantity) {
         super(name, category, price, quantity);
         this.setBreakable(true);
         this.setPerishable(false);
