+++ src/Inventory/Items/FragileItem.java
@@ -5,7 +5,7 @@ import Inventory.Types.Breakable;
 public class FragileItem extends AbstractItem {
 
 
-    public FragileItem(String name, String category, double price, double quantity) {
+    public FragileItem(String name, String category, double price, int quantity) {
         super(name, category, price, quantity);
         this.setBreakable(true);
         this.setPerishable(false);
